requirements.txt
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
pyTelegramBotAPI==4.3.1
kafka-python==2.0.2 
requests==2.26.0 
pyspark==3.5.2